Set paths back to PuppetLabs/Puppet on Windows
Changing the metadata unintentionally broke some things because changing company_id and product_id changes the paths where things live, where some modules expect things to be. This creates different variables for the directories so we can separate that from the ID itself. This will require more changes in the openvox-agent repo.