Setup "prepare release" and "release" actions
The `rake vox:tag` task is error prone as it push by default, without an
opportunity for a review, or updating a changelog.

Rely on the shared actions to prepare release, do release.  These expect
a version number but this project is special as the current date is the
version.  Current tags are barely readable in the format YYYYMMDDX where
X is a digit used to distinguish 2 releseas made the same day.

With this change, we move to the format YYYY-MM-DD-X which we hope is
more readable.  It is basically a form of [CalVer](https://calver.org/).

Because the UTC date is the same all over the world, and to avoid
mistakes, we do not allow the user to specify it when preparing / doing
releases.  The only drawback that we can see is that we can have
inconsistency if a release is prepared on one day, but the actual
release happen the following day.  It is not considered a critical thing
since we already have similar date inconsistencies in CHANGELOGs from
time to time for user facing projects, and this one is less problematic.