Wrap versionning logic in its own class
This project "version" is special as it is based on the current date
plus a trailing number to allow multiple releases on the same day.

When cutting a new release, to determine version to use, we must check
if the previous release was cut on the same day.  If so, we only want to
bumy hte trailing number, otherwise, we want to set the date to the
current date and reset the trailing number to 1.