Changes to enable building all MacOS types on any MacOS host
This makes several changes that enable us to build for all supported
MacOS versions on any kind of MacOS host running the latest version. It
sets some flags to tell clang to generate binaries compatible with
everything since MacOS 13. This also adds an osx-15-x86_64 platform, and
tweaks the build task to make sure you are running in a shell for the
architecture you are trying to build for. By using the 'arch' command to
open a shell for x86_64 before invoking the build rake task, we use
Rosetta 2 to allow us to build for x86_64 on an arm64 host just like we
would if we were building on an x86_64 host. This allows us to avoid any
of the many workarounds needed for cross compiling. While this method is
rather slow, it easier and potentially safer than all of the workarounds
needed for cross-compilation.

This also uses a platform string of macos-all-<arch> since we are now
able to build for all supported MacOS versions in one binary. This also
modifies the build action to allow overriding the vanagon branch and
to be able to use a "dev" branch of shared-actions for testing.
Unfortunately, the 'uses' parameter has to be a static string, which is
why we are limited to a specific branch named 'dev'.