Change windows-2019-x64 -> windows-all-x64
Because the single Windows agent should work for all supported modern
Windows OSes, this changes the name to windows-all-x64. This also adds
the site to the Cygwin bootstrapping setup.bat file, as this is needed
when running the installer for the very first time like we will be in
GitHub Actions.