Change version format to use periods
Because vanagon replaces all dashes in tags or git describes with periods, we end up with artifacts having different file names than the tag, and they end up being stored in a artifacts.voxpupuli.org in a directory with the tag that differs from the filename. While this isn't a huge deal, it makes automation more complicated. In order to standardize, this changes the tag format to use periods instead.