[CVE] remove nokogiri and dependencies
Nokogiri was used to bump performance on MacOS clients when parsing
plist files. Without it, CFPropertyList will fall back to rexml.
However, performance testing indicates that the boost is negligible.
This just removes it and reduces our security exposure.

This removes the libxml2 & libxslt libraries from the agent runtime, but
doesn't actually delete the components yet because Bolt uses them.