Prepare rubygem files for automation, rename bolt -> openbolt
This adds magic comments and formatting in preparation for a rake task
to update gems and their dependencies automatically. This also
simplifies the logic in some of them, removing paths for platforms we
don't intend to build and removing support for Ruby 2.7 in
agent-runtime-7.x for some components. This will be removed more
completely in a future update.