Add update_gems rake task
This rake task will inspect the data in each rubygems-*.rb file and use the RubyGems API to find if there are any new versions available, and update accordingly. It will also add missing or new dependencies, and if those are not yet added to the repo, it will create a component file for those.