Massive cleanup
This attempts to clean up the logic in as many configs as possible. Not
only has the code turned into a bit of spaghetti over the years, but it
contains a lot of dead code.

In general, the strategy here is to:
- Remove any platforms we will never build for (old).
- Keep platforms we don't currently build, but might in the future.
- In the future, we plan to move these runtime builds into the repos
  themselves (openvox/packaging and openbolt/packaging).
- After this, each branch will maintain its version of the packaging
  code.

Specifically, this commit:
- Removes anything to do with Ruby 2.7. If we need to go back to build
  7.x, we can make a branch from a prior commit.
- Removes anything to do with AIX 7.1 or SLES 11, any 32-bit platforms,
  and any other platforms we will never build for. Leave in AIX and
Solaris 10/11 code, as there is a nonzero chance we might look at these
platforms in the future.
- Reduces code branching and indirection by integrating several files
  that were previously shared by different versions of components. We
should not need to build more than one Ruby version for a particular
OpenVox branch, for example, so integrate _base-ruby.rb into the ruby
component.
- Removes old components no longer used.
- Integrates patches from mac_crosscompile branch. We aren't 100% sure
  we will be doing cross compilation for macos-all-x86_64, so attempted
to keep things as flexible as possible until then.
- Attempts to make code more readable and compact