Add openssl gem to Gemfile
In Ruby 3.2.9, the openssl gem is out of date. On MacOS with the latest OpenSSL, this breaks. This adds the gem to the gemfile so we don't break during builds. This doesn't affect the build itself, as it is due to OpenSSL 3.6 and we ship 3.0.

This doesn't quite work on Windows, where we have no choice but to use Ruby 3.2.2 in Cygwin.